refactor: tally
